% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBeta.R
\name{computeBeta}
\alias{computeBeta}
\title{Compute the parameter of Beta function given leaf angle distribution mearuements.}
\usage{
computeBeta(LeafAngles)
}
\arguments{
\item{LeafAngles}{The measurements of leaf angle distribution.}
}
\value{
The two parameters of Beta function given leaf angle distribution mearuements.
}
\description{
Compute the parameter of Beta function given leaf angle distribution mearuements.
}
\examples{
data(Pisek)
computeBeta(Pisek[[2]])
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
