% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGEllip.R
\name{computeGEllip}
\alias{computeGEllip}
\title{Compute the G value given lamda (the parameter of ellipsoidal function) and beam direction.}
\usage{
computeGEllip(lambda, theta)
}
\arguments{
\item{lambda}{The parameter of ellipsoidal function given leaf angle distribution mearuements.}

\item{theta}{The zenith angle of beam direction.}
}
\value{
The G value.
}
\description{
Compute the G value given lamda (the parameter of ellipsoidal function) and beam direction.
}
\examples{
computeGEllip(1.0, 30)

}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
