% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTrig.R
\name{computeTrig}
\alias{computeTrig}
\title{Compute the parameter of Verhoef's leaf angle distribution given leaf angle mearuements.}
\usage{
computeTrig(LeafAngles)
}
\arguments{
\item{LeafAngles}{The measurements of leaf angle distribution.}
}
\value{
The two parameters of Verhoef's leaf angle distribution.
}
\description{
Compute the parameter of Verhoef's leaf angle distribution given leaf angle distribution mearuements.
}
\examples{
data(Pisek)
computeTrig(Pisek[[2]])
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
