% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Falster.R
\docType{data}
\name{Falster}
\alias{Falster}
\title{Leaf angle distribution measurements provided by D.S. Falster}
\format{A list with 39 arrays where one array for tree specy name and 38 array of leaf angle distribution for each specy.}
\source{
Falster, D. and Westoby, M., 2003. Leaf size and angle vary widely across species: what consequences for light interception? New Phytol.158, 509-525.
}
\usage{
data(Falster)
}
\description{
Leaf angle distribution data measured in two sites of Ku-ringgai Chase National Park, Sydney, Australia
and provided by Dr. D.S. Falster of Macquarie University, Australia
}
\details{
Leaf angle distribution measurements provided by D.S. Falster
}
\examples{
data(Falster) 
Falster[1]
}
\author{
W.-M. Wang (wmwangATgmail.com)
}
\keyword{datasets}
