% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interEllip.R
\name{interEllip}
\alias{interEllip}
\title{Compute the fraction leaf area given leaf angle intervals (in degree) and the parameters of ellipsoidal function.}
\usage{
interEllip(AngleLower, AngleUpper, lambda)
}
\arguments{
\item{AngleLower}{The lower limit of leaf angle intervals in degree.}

\item{AngleUpper}{The upper limit of leaf angle intervals in degree.}

\item{lambda}{The parameter of ellipsoidal function.}
}
\value{
Compute the fraction leaf area of ellipsoidal function.
}
\description{
Compute the fraction leaf area given leaf angle intervals (in degree) and the parameters of ellipsoidal function.
}
\examples{
interEllip(40,50,1.1)
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
