% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interEllipR.R
\name{interEllipR}
\alias{interEllipR}
\title{Compute the fraction leaf area for rotated-ellipsoidal function given leaf angle intervals (in degree) and the parameters.}
\usage{
interEllipR(AngleLower, AngleUpper, lambdaR)
}
\arguments{
\item{AngleLower}{The lower limit of leaf angle intervals in degree.}

\item{AngleUpper}{The upper limit of leaf angle intervals in degree.}

\item{lambdaR}{The parameter of rotated-ellipsoidal function.}
}
\value{
Compute the fraction leaf area of rotated-ellipsoidal function.
}
\description{
Compute the fraction leaf area given leaf angle intervals (in degree) and the parameters of rotated-ellipsoidal function.
}
\examples{
interEllipR(40,50,1.1)
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
