% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interUniform.R
\name{interUniform}
\alias{interUniform}
\title{Compute the fraction leaf area of Uniform distribution given leaf angle intervals (in degree).}
\usage{
interUniform(AngleLower, AngleUpper)
}
\arguments{
\item{AngleLower}{The lower limit of leaf angle intervals in degree.}

\item{AngleUpper}{The upper limit of leaf angle intervals in degree.}
}
\value{
Compute the fraction leaf area of Uniform function.
}
\description{
Compute the fraction leaf area of Uniform distribution given leaf angle intervals (in degree).
}
\examples{
interUniform(40,50)
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
