% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LADdensity.R
\name{LADdensity}
\alias{LADdensity}
\title{Retrieve the density of leaf area given leaf angle measurements.}
\usage{
LADdensity(LeafAngles)
}
\arguments{
\item{LeafAngles}{The measurements of leaf angle distribution.}
}
\value{
Leaf area density given leaf angle intervals (0,10,20,30,40,50,60,70,80,90).
}
\description{
Retrieve the density of leaf area given leaf angle measurements.
}
\examples{
data(Falster) 
LADdensity(Falster[[2]])
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
