% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pisek.R
\docType{data}
\name{Pisek}
\alias{Pisek}
\title{Leaf angle distribution measurements provided by Jan Pisek}
\format{A list with 55 arrays where one array for tree specy name and 54 array of leaf angle distribution for each specy.}
\source{
Pisek, Jan, et al. "Is the spherical leaf inclination angle distribution a valid assumption for temperate and boreal broadleaf tree species?." 
Agricultural & Forest Meteorology 169.3(2013):186-194.
}
\usage{
data(Pisek)
}
\description{
Leaf angle distribution data provided by Dr. Tartu Observatory, Department of Remote Sensing, Estonia
}
\details{
Leaf angle distribution measurements provided by Jan Pisek
}
\examples{
data(Pisek)
Pisek[1]
}
\author{
W.-M. Wang (wmwangATgmail.com)
}
\keyword{datasets}
