% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdentsity.R
\name{plotdensity}
\alias{plotdensity}
\title{Plot the density of leaf area given leaf angle measurements.}
\usage{
plotdensity(LeafAngles, main = "Leaf angle distribution", type = "l",
  pch = 20)
}
\arguments{
\item{LeafAngles}{The measurements of leaf angle distribution.}

\item{main}{An overall title for the plot.}

\item{type}{The type of plot.}

\item{pch}{The symbol of plot.}
}
\description{
Plot the density of leaf area given leaf angle measurements.
}
\examples{
data(Pisek)
plotdensity(Pisek[[2]])
}
\references{
Wang, W. M., Li, Z. L., & Su, H. B. (2007). 
Comparison of leaf angle distribution functions: effects on extinction coefficient and 
fraction of sunlit foliage. Agricultural and Forest Meteorology, 143(1), 106-122.
}
\author{
Wei-Min Wang (wmwang AT gmail.com)
}
